+++ test/squint/compiler_test.cljs
@@ -7,6 +7,7 @@
    [clojure.test :as t :refer [async deftest is testing]]
    [promesa.core :as p]
    [squint.compiler :as squint]
+   [squint.dce-test]
    [squint.html-test]
    [squint.jsx-test]
    [squint.lazy-memory-test]
@@ -3286,4 +3287,5 @@ new Foo();")
                'squint.html-test
                'squint.math-test
                'squint.walk-test
-               'squint.multi-test))
+               'squint.multi-test
+               'squint.dce-test))